Financial Agent With Phidata/requirements.txt
phidata 
python-dotenv
yfinance
packaging
duckduckgo-search
googlesearch-python 
pycountry
fastapi
uvicorn
groq
openai